requirements.txt
torch>=2.0.0
numpy>=1.21.0
pandas>=1.3.0
tqdm>=4.65.0
python-chess>=1.9.0